+++ CHANGELOG.md
@@ -1,3 +1,8 @@
+## 0.0.35
+
+* Added a description about time zone handling to the README.
+* Added setOffset, setStartAfter, setEndBefore, setLimit method to QueryBuilder and RawQueryBuilder.
+
 ## 0.0.34
 
 * In search queries, sortObj is no longer required when using offset, startAfter, or endBefore. If not specified, the queries will be processed in the order they were added to the database.